requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
alabaster=0.7.12=py37_0
asn1crypto=0.24.0=py37_0
attrs=19.1.0=py37_1
babel=2.6.0=pypi_0
backcall=0.1.0=py37_0
beautifulsoup4=4.7.1=py37_1
blas=1.0=mkl
bleach=3.1.0=py37_0
bottleneck=1.2.1=py37h035aef0_1
bzip2=1.0.6=h14c3975_1002
ca-certificates=2019.5.15=0
cairo=1.14.12=h8948797_3
certifi=2019.6.16=py37_0
cffi=1.12.2=py37h2e261b9_1
chardet=3.0.4=py37_1
commonmark=0.8.1=py37_0
cryptography=2.6.1=py37h1ba5d50_0
cudatoolkit=10.0.130=0
cycler=0.10.0=py37_0
cymem=2.0.2=py37hfd86e86_0
cytoolz=0.9.0.1=py37h14c3975_1
dataclasses=0.6=py_0
dbus=1.13.6=h746ee38_0
decorator=4.4.0=py37_1
defusedxml=0.5.0=py37_1
dill=0.2.9=py37_0
docutils=0.14=pypi_0
entrypoints=0.3=py37_0
expat=2.2.6=he6710b0_0
fastprogress=0.1.21=py_0
ffmpeg=4.0.2=ha0c5888_2
fontconfig=2.13.0=h9420a91_0
freeglut=3.0.0=hf484d3e_1005
freetype=2.9.1=h8a8886c_1
future=0.17.1=py37_0
glib=2.56.2=hd408876_0
gmp=6.1.2=h6c8ec71_1
gnutls=3.5.19=h2a4e5f8_1
graphite2=1.3.13=hf484d3e_1000
gst-plugins-base=1.14.0=hbbd80ab_1
gstreamer=1.14.0=hb453b48_1
harfbuzz=1.9.0=he243708_1001
hdf5=1.10.2=hc401514_3
icu=58.2=h9c2bf20_1
idna=2.8=py37_0
imagesize=1.1.0=pypi_0
intel-openmp=2019.3=199
ipykernel=5.1.1=py37h24bf2e0_0
ipython=7.4.0=py37h39e3cac_0
ipython_genutils=0.2.0=py37_0
ipywidgets=7.4.2=py37_0
jasper=2.0.14=h07fcdf6_1
jedi=0.13.3=py37_0
jinja2=2.10.1=py37_0
jpeg=9b=h024ee3a_2
jsonschema=3.0.1=py37_0
jupyter=1.0.0=py37_7
jupyter_client=5.2.4=py37_0
jupyter_console=6.0.0=py37_0
jupyter_contrib_core=0.3.3=py_2
jupyter_contrib_nbextensions=0.5.1=py37_0
jupyter_core=4.4.0=py37_0
jupyter_highlight_selected_word=0.2.0=py37_1000
jupyter_latex_envs=1.4.4=py37_1000
jupyter_nbextensions_configurator=0.4.1=py37_0
kiwisolver=1.0.1=py37hf484d3e_0
libedit=3.1.20181209=hc058e9b_0
libffi=3.2.1=hd88cf55_4
libgcc-ng=8.2.0=hdf63c60_1
libgfortran=3.0.0=1
libgfortran-ng=7.3.0=hdf63c60_0
libglu=9.0.0=hf484d3e_1000
libiconv=1.15=h516909a_1005
libopencv=3.4.2=hb342d67_1
libpng=1.6.36=hbc83047_0
libsodium=1.0.16=h1bed415_0
libstdcxx-ng=8.2.0=hdf63c60_1
libtiff=4.0.10=h2733197_2
libuuid=1.0.3=h1bed415_2
libxcb=1.13=h1bed415_1
libxml2=2.9.9=he19cac6_0
libxslt=1.1.32=h4785a14_1002
lxml=4.3.3=py37h7ec2d77_0
m2r=0.2.1=py_0
markupsafe=1.1.1=py37h7b6447c_0
matplotlib=3.0.3=py37h5429711_0
mistune=0.8.4=py37h7b6447c_0
mkl=2018.0.3=1
mkl_fft=1.0.6=py37h7dd41cf_0
mkl_random=1.0.1=py37h4414c95_1
msgpack-numpy=0.4.3.2=py37_0
msgpack-python=0.6.1=py37hfd86e86_1
murmurhash=1.0.2=py37he6710b0_0
nbconvert=5.4.1=py37_3
nbformat=4.4.0=py37_0
ncurses=6.1=he6710b0_1
nettle=3.3=0
ninja=1.9.0=py37hfd86e86_0
notebook=5.7.8=py37_0
numexpr=2.6.8=py37hd89afb7_0
numpy=1.15.4=py37h1d66e8a_0
numpy-base=1.15.4=py37h81de0dd_0
nvidia-ml-py3=7.352.0=py_0
olefile=0.46=py37_0
opencv=3.4.2=py37h6fd60c2_1
openh264=1.8.0=hdbcaa40_1000
openssl=1.1.1c=h7b6447c_1
packaging=19.0=py37_0
pandas=0.24.2=py37he6710b0_0
pandoc=2.7.2=0
pandocfilters=1.4.2=py37_1
parso=0.5.0=py_0
pcre=8.43=he6710b0_0
pexpect=4.6.0=py37_0
pickleshare=0.7.5=py37_0
pillow=5.4.1=py37h34e0f95_0
pip=19.0.3=py37_0
pixman=0.38.0=h516909a_1003
plac=0.9.6=py37_0
preshed=2.0.1=py37he6710b0_0
prometheus_client=0.6.0=py37_0
prompt_toolkit=2.0.9=py37_0
ptyprocess=0.6.0=py37_0
py-opencv=3.4.2=py37hb342d67_1
pycparser=2.19=py37_0
pygments=2.3.1=py37_0
pyopenssl=19.0.0=py37_0
pyparsing=2.3.1=py37_0
pyqt=5.9.2=py37h05f1152_2
pyrsistent=0.14.11=py37h7b6447c_0
pysocks=1.6.8=py37_0
python=3.7.3=h0371630_0
python-dateutil=2.8.0=py37_0
pytorch=1.0.1=py3.7_cuda10.0.130_cudnn7.4.2_2
pytz=2018.9=py37_0
pyyaml=5.1=py37h7b6447c_0
pyzmq=18.0.0=py37he6710b0_0
qt=5.9.7=h5867ecd_1
qtconsole=4.4.3=py37_0
readline=7.0=h7b6447c_5
regex=2018.01.10=py37h14c3975_1000
scipy=1.1.0=py37hfa4b5c9_1
send2trash=1.5.0=py37_0
setuptools=40.8.0=py37_0
sip=4.19.8=py37hf484d3e_0
six=1.12.0=py37_0
snowballstemmer=1.2.1=pypi_0
soupsieve=1.8=py37_0
sphinxcontrib-applehelp=1.0.1=py_0
sphinxcontrib-devhelp=1.0.1=py_0
sphinxcontrib-htmlhelp=1.0.2=py_0
sphinxcontrib-jsmath=1.0.1=py_0
sphinxcontrib-qthelp=1.0.2=py_0
sphinxcontrib-serializinghtml=1.1.3=py_0
sqlite=3.27.2=h7b6447c_0
terminado=0.8.1=py37_1
testpath=0.4.2=py37_0
thinc=6.12.0=py37h4989274_0
tk=8.6.8=hbc83047_0
toolz=0.9.0=py37_0
torchvision=0.2.2=py_3
tornado=6.0.2=py37h7b6447c_0
tqdm=4.31.1=py37_1
traitlets=4.3.2=py37_0
typing=3.6.4=py37_0
ujson=1.35=py37h14c3975_0
urllib3=1.25.3=py37_0
wcwidth=0.1.7=py37_0
webencodings=0.5.1=py37_1
wheel=0.33.1=py37_0
widgetsnbextension=3.4.2=py37_0
wrapt=1.10.11=py37h14c3975_2
x264=1!152.20180806=h14c3975_0
xorg-fixesproto=5.0=h14c3975_1002
xorg-inputproto=2.3.2=h14c3975_1002
xorg-kbproto=1.0.7=h14c3975_1002
xorg-libx11=1.6.7=h14c3975_1000
xorg-libxau=1.0.9=h14c3975_0
xorg-libxext=1.3.4=h516909a_0
xorg-libxfixes=5.0.3=h516909a_1004
xorg-libxi=1.7.10=h516909a_0
xorg-xextproto=7.3.0=h14c3975_1002
xorg-xproto=7.0.31=h14c3975_1007
xz=5.2.4=h14c3975_4
yaml=0.1.7=had09818_2
zeromq=4.3.1=he6710b0_3
zlib=1.2.11=h7b6447c_3
zstd=1.3.7=h0b5b093_0
